docs: a separate public config with GA and improved search.
Adding GA tracking to the docs, but to avoid privacy issues in a local
build, there's now a separate public config. Additionally, for the
public version the search downloads a binary data (which is smaller, but
won't work locally), plus also provides a helpful link to Google search
if nothing is found.